+++ libft/ft_printf/srcs/write_float_3.c
@@ -6,7 +6,7 @@
 /*   By: ehelmine <ehelmine@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/26 21:04:23 by ehelmine          #+#    #+#             */
-/*   Updated: 2021/09/22 15:41:45 by ehelmine         ###   ########.fr       */
+/*   Updated: 2022/02/08 16:10:54 by ehelmine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ static int	write_float_output_2(t_val *all, int i)
 		all->str[i++] = '.';
 	if (all->orig_precision != 0)
 	{
-		x = ft_strlen(all->end_str);
+		x = (int)ft_strlen(all->end_str);
 		while (x < all->orig_precision)
 		{
 			all->str[i++] = '0';
@@ -81,7 +81,7 @@ int	write_float_output(t_val *all)
 	if (all->negative_val < 1 && all->plus_flag == 1 && all->space_flag < 1)
 		all->str[all->y++] = '+';
 	all->y = write_float_output_2(all, all->y);
-	all->len = ft_strlen(all->str);
+	all->len = (int)ft_strlen(all->str);
 	if (all->len < all->width)
 		write_float_output_3(all, all->y);
 	all->output_len += all->len;
@@ -98,7 +98,7 @@ int	write_float_9(t_val *all)
 	all->end_str = ft_itoa(all->end_i);
 	if (all->end_str == NULL)
 		return (-1);
-	all->y = ft_strlen(all->end_str) - 1;
+	all->y = (int)ft_strlen(all->end_str) - 1;
 	while (all->y >= 0 && all->end_str[all->y] == '9')
 		all->y--;
 	all->end_str[all->y] = all->end_str[all->y] + 1;
