+++ main.c
@@ -6,7 +6,7 @@
 /*   By: ehelmine <ehelmine@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/06 09:58:56 by ehelmine          #+#    #+#             */
-/*   Updated: 2022/02/07 18:44:10 by ehelmine         ###   ########.fr       */
+/*   Updated: 2022/02/08 16:08:06 by ehelmine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ static void	args_to_struct(int argc, char **argv, t_select *data)
 	x = 0;
 	while (i < argc && x < MAX_INPUT_LEN - 1)
 	{
-		if (ft_strlen(argv[i]) >= MAX_INPUT_LEN)
+		if ((int)ft_strlen(argv[i]) >= MAX_INPUT_LEN)
 		{
 			write(STDERR_FILENO, "Argument too long, max length is 2048\n", 38);
 			exit (EXIT_FAILURE);
